requirements.txt
telegram==0.0.1
Flask==0.12.1
transitions==0.5.0
pygraphviz==1.4rc1
python-telegram-bot==5.3.0
facebook-sdk==2.0.0
